# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: GenPos — XiaoHongShu AI Ads Workspace API
  version: 1.0.0
  description: |
    Production-grade, multi-tenant, China-first XiaoHongShu AI creative operating
    system API. Enables merchants to manage products, generate XiaoHongShu-native
    note packages via AI agent teams, review and export content for 笔记, 聚光,
    and 蒲公英 surfaces, and track performance analytics.

    ## Authentication
    All endpoints require a valid Bearer JWT. The token embeds `tenant_id`,
    `user_id`, and RBAC roles. Obtain tokens via the Identity & Tenant Service
    (`/auth/token`).

    ## Pagination
    List endpoints support `limit` / `offset` query parameters. Default `limit`
    is 20, max 100. Responses include a `pagination` object with `total`,
    `limit`, `offset`, and `has_more`.

    ## Multi-Tenancy
    Tenant isolation is enforced at the gateway. The JWT `tenant_id` claim scopes
    all data access — callers never need to pass a tenant parameter explicitly.
  contact:
    name: GenPos Engineering
    email: engineering@genpos.ai
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: https://api.genpos.ai/api/v1
    description: Production
  - url: https://staging-api.genpos.ai/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /products:
    post:
      operationId: createProduct
      tags: [Products]
      summary: Create product
      description: |
        Add a new product to the merchant's catalog. Products represent
        individual SKUs or product lines that will receive generated
        XiaoHongShu content.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /products/{id}:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: getProduct
      tags: [Products]
      summary: Get product
      description: Retrieve a product by ID, including its current status and metadata.
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateProduct
      tags: [Products]
      summary: Update product
      description: Partially update a product's profile. Only supplied fields are changed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /merchants/{merchant_id}/products:
    parameters:
      - $ref: '#/components/parameters/MerchantIdPath'
    get:
      operationId: listMerchantProducts
      tags: [Products]
      summary: List merchant products
      description: |
        List all products belonging to a merchant. Supports pagination and
        optional status filtering.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
          description: Filter by product status
        - name: category
          in: query
          schema:
            type: string
          description: Filter by product category
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # =========================================================================
  # ASSET PACK ENDPOINTS
  # =========================================================================
...
  /asset-packs/{id}/assets:
    parameters:
      - $ref: '#/components/parameters/AssetPackId'
    post:
      operationId: uploadAsset
      tags: [Asset Packs]
      summary: Upload asset to pack
      description: |
        Upload a new asset (packshot, cutout, logo, or packaging reference)
        to an asset pack. The file is stored in object storage and metadata
        is recorded. The asset starts in `draft` approval status.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, type]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, WebP)
                type:
                  $ref: '#/components/schemas/AssetType'
                product_id:
                  type: string
                  format: uuid
                  description: Optional product association
                metadata:
                  type: string
                  description: JSON-encoded metadata object
      responses:
        '201':
          description: Asset uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: listAssetPackAssets
      tags: [Asset Packs]
      summary: List assets in pack
      description: List all assets within an asset pack with optional type filtering.
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AssetType'
          description: Filter by asset type
        - name: approval_status
          in: query
          schema:
            $ref: '#/components/schemas/ApprovalStatus'
          description: Filter by approval status
      responses:
        '200':
          description: Paginated asset list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /products/{id}/performance:
    parameters:
      - $ref: '#/components/parameters/ProductId'
    get:
      operationId: getProductPerformance
      tags: [Analytics]
      summary: Product performance
      description: |
        Retrieve aggregated performance metrics for a product across all
        its published note packages. Supports date range filtering and
        breakdown by source mode or objective.
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Start of the reporting period (inclusive)
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
          description: End of the reporting period (inclusive)
        - name: breakdown_by
          in: query
          schema:
            type: string
            enum: [source_mode, objective, style_family, day]
          description: Dimension to break down metrics by
      responses:
        '200':
          description: Product performance metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPerformanceReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /note-packages/{id}/performance:
    parameters:
      - $ref: '#/components/parameters/NotePackageId'
    get:
      operationId: getNotePackagePerformance
      tags: [Analytics]
      summary: Note package performance
      description: |
        Retrieve daily performance metrics for a specific note package.
        Includes engagement, conversion, cost, and computed fatigue score.
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Start of the reporting period
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: End of the reporting period
      responses:
        '200':
          description: Note package performance metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotePackagePerformanceReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # =========================================================================
  # PERSONA ENDPOINTS
  # =========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256-signed JWT. Claims include `sub` (user ID), `tenant_id`,
        `roles` (array of RBAC roles: owner, admin, editor, viewer), and
        standard `exp`/`iat`/`iss` fields. Tokens are issued by the Identity
        & Tenant Service.

  # -------------------------------------------------------------------------
  # PARAMETERS
  # -------------------------------------------------------------------------
  parameters:
...
    MerchantIdPath:
      name: merchant_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Merchant UUID
    ProductId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Product UUID
    AssetPackId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Asset pack UUID
    NotePackageId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Note package UUID
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip

  # -------------------------------------------------------------------------
  # REUSABLE RESPONSES
  # -------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Invalid request syntax or malformed parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, expired, or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Request semantically invalid (validation failure)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # -------------------------------------------------------------------------
  # SCHEMAS
  # -------------------------------------------------------------------------
  schemas:
...
    Pagination:
      type: object
      required: [total, limit, offset, has_more]
      properties:
        total:
          type: integer
          description: Total number of matching records
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              examples: ['MERCHANT_NOT_FOUND']
            message:
              type: string
              description: Human-readable error description
            request_id:
              type: string
              format: uuid
              description: Correlation ID for debugging
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              examples: ['VALIDATION_ERROR']
            message:
              type: string
            request_id:
              type: string
              format: uuid
            details:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    description: JSON path to the invalid field
                  message:
                    type: string
                  type:
                    type: string
                    description: Validation rule that failed

    # -----------------------------------------------------------------------
    # ENUMS
    # -----------------------------------------------------------------------
...
    ProductStatus:
      type: string
      enum: [active, inactive, archived]
...
    AssetType:
      type: string
      enum: [packshot, cutout, logo, packaging_ref, hero, supplementary]
    ApprovalStatus:
      type: string
      enum: [draft, pending_review, approved, archived]
...
    Objective:
      type: string
      enum: [seeding, conversion, brand_awareness, event_traffic]
...
    ProductCreate:
      type: object
      required: [merchant_id, name, category]
      properties:
        merchant_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        category:
          type: string
          maxLength: 100
          description: Product category (e.g., 防晒霜, 面膜, 零食)
        description:
          type: string
          maxLength: 2000
        primary_objective:
          $ref: '#/components/schemas/Objective'
        selling_points:
          type: array
          items:
            type: string
          description: Key selling points for this product
        target_audiences:
          type: array
          items:
            type: string
          description: Target audience descriptions
        price:
          type: number
          format: float
          description: Product price in CNY
        price_display:
          type: string
          description: Formatted price string for display (e.g., ¥99)
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        category:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        status:
          $ref: '#/components/schemas/ProductStatus'
        primary_objective:
          $ref: '#/components/schemas/Objective'
        selling_points:
          type: array
          items:
            type: string
        target_audiences:
          type: array
          items:
            type: string
        price:
          type: number
          format: float
        price_display:
          type: string
    Product:
      type: object
      required: [id, merchant_id, name, category, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProductStatus'
        primary_objective:
          $ref: '#/components/schemas/Objective'
          nullable: true
        selling_points:
          type: array
          items:
            type: string
        target_audiences:
          type: array
          items:
            type: string
        price:
          type: number
          format: float
          nullable: true
        price_display:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # -----------------------------------------------------------------------
    # ASSET PACK
    # -----------------------------------------------------------------------
...
    Asset:
      type: object
      required: [id, asset_pack_id, type, storage_url, approval_status, created_at]
      properties:
        id:
          type: string
          format: uuid
        asset_pack_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
          nullable: true
        type:
          $ref: '#/components/schemas/AssetType'
        storage_url:
          type: string
          format: uri
          description: CDN or object-storage URL
        checksum:
          type: string
          description: SHA-256 checksum of the file
        width:
          type: integer
          description: Image width in pixels
        height:
          type: integer
          description: Image height in pixels
        file_size_bytes:
          type: integer
        mime_type:
          type: string
          examples: ['image/jpeg', 'image/png', 'image/webp']
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary metadata (color profile, DPI, etc.)
        approval_status:
          $ref: '#/components/schemas/ApprovalStatus'
        created_at:
          type: string
          format: date-time
    AssetList:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # -----------------------------------------------------------------------
    # GENERATION JOB / TASK
    # -----------------------------------------------------------------------
...
    PerformanceMetrics:
      type: object
      required: [id, note_package_id, date, impressions, clicks, saves, comments]
      properties:
        id:
          type: string
          format: uuid
        note_package_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        impressions:
          type: integer
          minimum: 0
        clicks:
          type: integer
          minimum: 0
        saves:
          type: integer
          minimum: 0
          description: 收藏 count
        comments:
          type: integer
          minimum: 0
        cost:
          type: number
          format: float
          minimum: 0
          nullable: true
          description: Ad spend in CNY (for paid content)
        conversions:
          type: integer
          minimum: 0
          nullable: true
        revenue:
          type: number
          format: float
          minimum: 0
          nullable: true
          description: Attributed revenue in CNY
        ctr:
          type: number
          format: float
          nullable: true
          description: Computed click-through rate
        save_rate:
          type: number
          format: float
          nullable: true
          description: Computed save rate (saves / impressions)
        fatigue_score:
          type: number
          format: float
          nullable: true
          minimum: 0
          maximum: 1
          description: Computed fatigue score (0 = fresh, 1 = saturated)
...
    ProductPerformanceReport:
      type: object
      required: [product_id, start_date, end_date, totals]
      properties:
        product_id:
          type: string
          format: uuid
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        totals:
          $ref: '#/components/schemas/PerformanceSummary'
        breakdown:
          type: array
          items:
            type: object
            required: [dimension_value, metrics]
            properties:
              dimension_value:
                type: string
              metrics:
                $ref: '#/components/schemas/PerformanceSummary'
        top_note_packages:
          type: array
          items:
            type: object
            required: [note_package_id, ranking_score, metrics]
            properties:
              note_package_id:
                type: string
                format: uuid
              ranking_score:
                type: number
                format: float
              metrics:
                $ref: '#/components/schemas/PerformanceSummary'
    NotePackagePerformanceReport:
      type: object
      required: [note_package_id, daily_metrics]
      properties:
        note_package_id:
          type: string
          format: uuid
        cumulative:
          $ref: '#/components/schemas/PerformanceSummary'
        daily_metrics:
          type: array
          items:
            $ref: '#/components/schemas/PerformanceMetrics'
        fatigue_trend:
          type: array
          items:
            type: object
            required: [date, score]
            properties:
              date:
                type: string
                format: date
              score:
                type: number
                format: float
    PerformanceSummary:
      type: object
      required: [impressions, clicks, saves, comments]
      properties:
        impressions:
          type: integer
        clicks:
          type: integer
        saves:
          type: integer
        comments:
          type: integer
        conversions:
          type: integer
          nullable: true
        cost:
          type: number
          format: float
          nullable: true
        revenue:
          type: number
          format: float
          nullable: true
        ctr:
          type: number
          format: float
          nullable: true
        save_rate:
          type: number
          format: float
          nullable: true
        cost_per_conversion:
          type: number
          format: float
          nullable: true
        roas:
          type: number
          format: float
          nullable: true
          description: Return on ad spend

    # -----------------------------------------------------------------------
    # PERSONA
    # -----------------------------------------------------------------------
...